+++ Assets/Scripts/Trail Making/TrailLevel3.cs
@@ -13,7 +13,7 @@ public class TrailLevel3 : MonoBehaviour
 
     public GameObject trail31;
     public GameObject trail32;
-    public int levelThreeScore = 0;
+    public float levelThreeScore = 0;
     public int mistakes = 0;
     public float levelThreeAccuracy = 0;
     public  float correctCounter = 0;
